AnimeJS removal, logic for resizing/timing
In order to facilitate resizing this commit adds in the logic for
updating the filter on OBS and then triggering it to the new size.

Additionally it removes AnimeJS and adds in nodemon and adds in a `npm
run dev` command to start up a dev server that will reload as you make
changes.