Add minimum/maximum commands and variables
In order to allow us to tweak min/max sizes on the fly this commit
introduces two new command URL's of `/min` and `max` that both take an
`?amount=` query and set that as the new minimum or maximum.

Additionally we now can set them at the beginning via MINIMUM_SIZE and
MAXIMUM_SIZE environment variables.